cloudkms: Move Signer to separate file
